Refactor and add all int and uint fields
